02_mqtt-cluster/requirements.txt
wheel
paho-mqtt==1.6.1
logger
pytest